code/lambdas/action-lambda/requirements.txt
boto3==1.34.57
llama-index==0.10.13
llama-index-embeddings-bedrock==0.1.3 
llama-index-llms-bedrock==0.1.3
sqlalchemy==2.0.23
PyAthena[SQLAlchemy]
openai
vanna
pandas
opensearch-py
requests_aws4auth
opensearch-dsl
langchain-community
langchain-huggingface
matplotlib
plotly
pyyaml